rename: ResultDetail 컴포넌트 result/creator 폴더로 이동
