+++ src/pages/CreatorResult.tsx
@@ -7,7 +7,7 @@ import useNavbar from '@/hooks/useNavbar'
 import Navbar from '@/components/form/result/Navbar'
 import SignOut from '@/components/shared/SignOut'
 import { useGetAlbumInfo } from '@/hooks/useGetAlbumInfo'
-import ResultDetail from '@/components/form/result/ResultDetail'
+import ResultDetail from '@/components/form/result/creator/ResultDetail'
 import { useGetVoteCount } from '@/hooks/useGetVoteCount'
 import { Text } from '@/components/shared/text'
